+++ functions/src/tax/index.ts
@@ -1,0 +1,3 @@
+export * from './types.js'
+export * from './taxRepository.js'
+export * from './taxResolver.js'